+++ source/src/Forest/Forest.cpp
@@ -44,11 +44,11 @@
 #include "DataFloat.h"
 
 Forest::Forest() :
-    verbose_out(0), num_trees(DEFAULT_NUM_TREE), mtry(0), min_node_size(0), num_variables(0), num_independent_variables(
-        0), seed(0), dependent_varID(0), num_samples(0), prediction_mode(false), memory_mode(MEM_DOUBLE), sample_with_replacement(
-        true), memory_saving_splitting(false), splitrule(DEFAULT_SPLITRULE), predict_all(false), keep_inbag(false), num_threads(
-        DEFAULT_NUM_THREADS), data(0), overall_prediction_error(0), importance_mode(DEFAULT_IMPORTANCE_MODE), progress(
-        0) {
+verbose_out(0), num_trees(DEFAULT_NUM_TREE), mtry(0), min_node_size(0), num_variables(0), num_independent_variables(
+    0), seed(0), dependent_varID(0), num_samples(0), prediction_mode(false), memory_mode(MEM_DOUBLE), sample_with_replacement(
+    true), memory_saving_splitting(false), splitrule(DEFAULT_SPLITRULE), predict_all(false), keep_inbag(false), sample_fraction(1), num_threads(
+    DEFAULT_NUM_THREADS), data(0), overall_prediction_error(0), importance_mode(DEFAULT_IMPORTANCE_MODE), progress(
+    0) {
 }
 
 Forest::~Forest() {
@@ -62,7 +62,8 @@ void Forest::initCpp(std::string dependent_variable_name, MemoryMode memory_mode
     std::string load_forest_filename, ImportanceMode importance_mode, uint min_node_size,
     std::string split_select_weights_file, std::vector<std::string>& always_split_variable_names,
     std::string status_variable_name, bool sample_with_replacement, std::vector<std::string>& unordered_variable_names,
-    bool memory_saving_splitting, SplitRule splitrule, std::string case_weights_file, bool predict_all) {
+    bool memory_saving_splitting, SplitRule splitrule, std::string case_weights_file, bool predict_all,
+    double sample_fraction) {
 
   this->verbose_out = verbose_out;
 
@@ -96,7 +97,7 @@ void Forest::initCpp(std::string dependent_variable_name, MemoryMode memory_mode
   // Call other init function
   init(dependent_variable_name, memory_mode, data, mtry, output_prefix, num_trees, seed, num_threads, importance_mode,
       min_node_size, status_variable_name, prediction_mode, sample_with_replacement, unordered_variable_names,
-      memory_saving_splitting, splitrule, predict_all);
+      memory_saving_splitting, splitrule, predict_all, sample_fraction);
 
   if (prediction_mode) {
     loadFromFile(load_forest_filename);
@@ -140,14 +141,14 @@ void Forest::initR(std::string dependent_variable_name, Data* input_data, uint m
     std::vector<std::vector<double>>& split_select_weights, std::vector<std::string>& always_split_variable_names,
     std::string status_variable_name, bool prediction_mode, bool sample_with_replacement,
     std::vector<std::string>& unordered_variable_names, bool memory_saving_splitting, SplitRule splitrule,
-    std::vector<double>& case_weights, bool predict_all, bool keep_inbag) {
+    std::vector<double>& case_weights, bool predict_all, bool keep_inbag, double sample_fraction) {
 
   this->verbose_out = verbose_out;
 
   // Call other init function
   init(dependent_variable_name, MEM_DOUBLE, input_data, mtry, "", num_trees, seed, num_threads, importance_mode,
       min_node_size, status_variable_name, prediction_mode, sample_with_replacement, unordered_variable_names,
-      memory_saving_splitting, splitrule, predict_all);
+      memory_saving_splitting, splitrule, predict_all, sample_fraction);
 
   // Set variables to be always considered for splitting
   if (!always_split_variable_names.empty()) {
@@ -175,7 +176,7 @@ void Forest::init(std::string dependent_variable_name, MemoryMode memory_mode, D
     std::string output_prefix, uint num_trees, uint seed, uint num_threads, ImportanceMode importance_mode,
     uint min_node_size, std::string status_variable_name, bool prediction_mode, bool sample_with_replacement,
     std::vector<std::string>& unordered_variable_names, bool memory_saving_splitting, SplitRule splitrule,
-    bool predict_all) {
+    bool predict_all, double sample_fraction) {
 
   // Initialize data with memmode
   this->data = input_data;
@@ -212,6 +213,7 @@ void Forest::init(std::string dependent_variable_name, MemoryMode memory_mode, D
   this->memory_saving_splitting = memory_saving_splitting;
   this->splitrule = splitrule;
   this->predict_all = predict_all;
+  this->sample_fraction = sample_fraction;
 
   // Set number of samples and variables
   num_samples = data->getNumRows();
@@ -246,6 +248,11 @@ void Forest::init(std::string dependent_variable_name, MemoryMode memory_mode, D
   // Check if mtry is in valid range
   if (this->mtry > num_variables - 1) {
     throw std::runtime_error("mtry can not be larger than number of variables in data.");
+  }
+
+  // Check if any observations samples
+  if ((size_t) num_samples * sample_fraction < 1) {
+    throw std::runtime_error("sample_fraction too small, no observations sampled.");
   }
 }
 
@@ -399,7 +406,7 @@ void Forest::grow() {
 
     trees[i]->init(data, mtry, dependent_varID, num_samples, tree_seed, &deterministic_varIDs, &split_select_varIDs,
         tree_split_select_weights, importance_mode, min_node_size, &no_split_variables, sample_with_replacement,
-        &is_ordered_variable, memory_saving_splitting, splitrule, &case_weights, keep_inbag);
+        &is_ordered_variable, memory_saving_splitting, splitrule, &case_weights, keep_inbag, sample_fraction);
   }
 
   // Init variable importance