+++ source/src/main.cpp
@@ -89,7 +89,7 @@ int main(int argc, char **argv) {
         arg_handler.outprefix, arg_handler.ntree, verbose_out, arg_handler.seed, arg_handler.nthreads,
         arg_handler.predict, arg_handler.impmeasure, arg_handler.targetpartitionsize, arg_handler.splitweights,
         arg_handler.alwayssplitvars, arg_handler.statusvarname, arg_handler.replace, arg_handler.catvars,
-        arg_handler.savemem, arg_handler.splitrule, arg_handler.caseweights, arg_handler.predall);
+        arg_handler.savemem, arg_handler.splitrule, arg_handler.caseweights, arg_handler.predall, arg_handler.fraction);
     forest->run(true);
     if (arg_handler.write) {
       forest->saveToFile();